Fixing some small loop bugs.
I do not want to fill this out... So I will just fill this out with some random jargon..... jargeon.... or however you spell that word. Anyway I fixed the thing with the loop and some of the dialog in the email. And that should wrap up this small update.